Add gift cards methods

+++ src/main/java/crowstream/microservice/payment/application/services/GiftCardServiceImpl.java
@@ -5,7 +5,7 @@ import crowstream.microservice.payment.application.ports.GiftCardPersistencePort
 import crowstream.microservice.payment.application.ports.GiftCardServicePort;
 
 public class GiftCardServiceImpl implements GiftCardServicePort {
-    private GiftCardPersistencePort persistencePort;
+    private final GiftCardPersistencePort persistencePort;
 
     public GiftCardServiceImpl(GiftCardPersistencePort persistencePort) {
         this.persistencePort = persistencePort;
@@ -13,6 +13,9 @@ public class GiftCardServiceImpl implements GiftCardServicePort {
 
     @Override
     public GiftCard addGiftCard(GiftCard card) {
+        if (this.persistencePort.existsByCardCode(card.getCardCode())) {
+            return null;
+        }
         return this.persistencePort.save(card);
     }
 
@@ -22,12 +25,23 @@ public class GiftCardServiceImpl implements GiftCardServicePort {
     }
 
     @Override
+    public GiftCard getGiftCardByCardCode(String cardCode) {
+        return this.persistencePort.findByCardCode(cardCode);
+    }
+
+    @Override
     public GiftCard updateGiftCard(GiftCard card) {
+        if (!this.persistencePort.existsByCardCode(card.getCardCode())) {
+            return null;
+        }
         return this.persistencePort.update(card);
     }
 
     @Override
     public void deleteGiftCardById(Long id) {
+        if (this.persistencePort.findById(id) == null) {
+            return;
+        }
         this.persistencePort.deleteById(id);
     }
 }
